+++ odoo/local-src/specific_sale/models/__init__.py
@@ -1,3 +1,4 @@
 from . import company
 from . import sale
 from . import sale_order_line
+from . import account_invoice